New project format benchmarks.

+++ Gu.Reactive.Benchmarks/Program.cs
@@ -1,19 +1,16 @@
 ﻿namespace Gu.Reactive.Benchmarks
 {
+    using System;
     using System.Collections.Generic;
-    using System.Diagnostics.CodeAnalysis;
     using System.IO;
-
     using BenchmarkDotNet.Reports;
     using BenchmarkDotNet.Running;
 
-    [SuppressMessage("ReSharper", "UnusedMember.Local")]
-    [SuppressMessage("ReSharper", "UnusedParameter.Local")]
     public class Program
     {
-        //// ReSharper disable PossibleNullReferenceException
-        private static readonly string DestinationDirectory = Path.Combine(new DirectoryInfo(Directory.GetCurrentDirectory()).Parent.Parent.FullName, "Benchmarks");
-        //// ReSharper restore PossibleNullReferenceException
+        private static readonly string ProjectDirectory = Directory.GetCurrentDirectory();
+
+        private static string ArtifactsDirectory { get; } = Path.Combine(ProjectDirectory, "BenchmarkDotNet.Artifacts", "results");
 
         public static void Main()
         {
@@ -38,13 +35,14 @@
 
         private static void CopyResult(string name)
         {
-#if DEBUG
-#else
-            var sourceFileName = Path.Combine(Directory.GetCurrentDirectory(), "BenchmarkDotNet.Artifacts", "results", name + "-report-github.md");
-            Directory.CreateDirectory(DestinationDirectory);
-            var destinationFileName = Path.Combine(DestinationDirectory, name + ".md");
-            File.Copy(sourceFileName, destinationFileName, overwrite: true);
-#endif
+            Console.WriteLine($"DestinationDirectory: {ProjectDirectory}");
+            if (Directory.Exists(ProjectDirectory))
+            {
+                var sourceFileName = Path.Combine(ArtifactsDirectory, name + "-report-github.md");
+                var destinationFileName = Path.Combine(ProjectDirectory, "Benchmarks", name + ".md");
+                Console.WriteLine($"Copy: {sourceFileName} -> {destinationFileName}");
+                File.Copy(sourceFileName, destinationFileName, overwrite: true);
+            }
         }
     }
 }
